# Conflicts:
#	amoginarium/logic/_run_logic.py
#	amoginarium/logic/entities/__init__.py
#	amoginarium/logic/entities/_base/_base_entities/_base_logic_entity.py
#	amoginarium/logic/entities/_base/_debug/_debug_polygon.py
#	amoginarium/logic/entities/_weaponry/templates/_sensors/_base_sensor.py
#	amoginarium/logic/entities/_weaponry/templates/_turrets/_base_turret.py
#	amoginarium/logic/entities/_weaponry/templates/_turrets/_calculated_rideable_turret.py
#	setup.py

Took 6 seconds

+++ amoginarium/logic/entities/_base/_base_entities/_base_logic_entity.py
@@ -4,10 +4,10 @@ Defines BaseLogicEntity.
 Defines the most basic logic entity structure.
 Includes hierarchy management, lifecycle hooks, and bitwise buffer access.
 
-Path: amoginarium/logic/entities/_base/_base_entities/_base_logic_entity.py
-Project: amoginarium
-Created: 28.03.2026
-Authors: Nilusink, LukasKrah
+| ``Path``: amoginarium/logic/entities/_base/_base_entities/_base_logic_entity.py
+| ``Project``: amoginarium
+| ``Created``: 28.03.2026
+| ``Authors``: Nilusink, LukasKrah
 """
 
 from __future__ import annotations
@@ -23,33 +23,9 @@ if tp.TYPE_CHECKING:
     from ctypes import Array
     from types import EllipsisType
 
-    from amoginarium.shared import base_entity_t
+    from amoginarium.shared import base_entity_t, EntityChildViable, MurderViable
 
     from .._groups import LogicGroup
-
-
-class EntityChildViable(tp.Protocol):
-    """
-    Minimum requirements for an object to be assigned as the child of a logic entity.
-    """
-
-    def update(self, delta: float) -> None:
-        """
-        Update function.
-
-        :param delta: Tme since the last update
-        """
-
-    def kill(self) -> None:
-        """Clean up and terminate the child."""
-
-
-class MurderViable(tp.Protocol):
-    """Can kill someone."""
-
-    @property
-    def parent(self) -> tp.Any:
-        """Parent."""
 
 
 class BaseLogicEntity(BaseLogicEntityLike):
@@ -216,25 +192,44 @@ class BaseLogicEntity(BaseLogicEntityLike):
                 group.remove(self)
                 self.__groups.remove(group)
 
-    # noinspection PyUnusedLocal
+    # noinspection PyUnusedLocal, PyMethodMayBeStatic
+    def _before_kill(  # noqa: PLR6301
+        self,
+        *,
+        killed_by: MurderViable | EllipsisType = ...,  # noqa: ARG002
+        kill_children: bool = True,  # noqa: ARG002
+    ) -> bool:
+        """
+        Whether the entity can be killed. Called before _kill.
+
+        :param killed_by: Who killed this entity
+        :param kill_children: Whether to kill children as well recursively
+        :return: Whether the entity kill is accepted
+        """
+        return True
+
     def _kill(
         self,
+        *,
         killed_by: MurderViable | EllipsisType = ...,  # noqa: ARG002
+        kill_children: bool = True,
     ) -> None:
         """
         Kill the entity and all its children.
 
         :param killed_by: Who killed this entity
+        :param kill_children: Whether to kill children as well recursively
         """
         # kill children first
-        for child in self._children:
-            child.kill()
+        if kill_children:
+            for child in self._children:
+                child.kill()
 
         # commit suicide
         for group in self.__groups:
             group.remove(self)
 
-        self._set_bit("flags", 0, False)  # set alive # noqa: FBT003
+        self._set_bit("flags", 0, False)  # set alive  # noqa: FBT003
         ENTITY_COUNTER.pop_id(self.__id)
 
         self.__groups.clear()
@@ -242,16 +237,55 @@ class BaseLogicEntity(BaseLogicEntityLike):
         # add to dead
         Dead.add(self)
 
+    def _after_kill(
+        self,
+        *,
+        killed_by: MurderViable | EllipsisType = ...,
+        kill_children: bool = True,
+        killed: bool = True,
+    ) -> None:
+        """
+        Reaction at the end of kill no matter if the kill was accepted or not.
+
+        :param killed_by: Who killed this entity
+        :param kill_children: Whether to kill children as well recursively
+        :param killed: Whether the entity kill was accepted or not
+        """
+
     @tp.final
-    def kill(self, killed_by: MurderViable | EllipsisType = ...) -> None:
+    def kill(
+        self,
+        *,
+        killed_by: MurderViable | EllipsisType = ...,
+        kill_children: bool = True,
+        force_kill: bool = False,
+    ) -> bool | None:
         """
         Kill the entity and all its children.
 
         :param killed_by: Who killed this entity
+        :param kill_children: Whether to kill children as well as recursively
+        :param force_kill: Whether to kill even if before kill returns False
+        :return: Whether the entity wa0s killed or not. May be denied by _before_kill.
+            None if the entity is already dead.
         """
         if self._alive:
-            self._alive = False
-            self._kill(killed_by)
+            killed: bool = False
+
+            kill_entity: bool | None = self._before_kill(
+                killed_by=killed_by, kill_children=kill_children
+            )
+
+            if force_kill or kill_entity is True or kill_entity is None:
+                self._alive = False
+                self._kill(killed_by=killed_by, kill_children=kill_children)
+                killed = True
+
+            self._after_kill(
+                killed_by=killed_by, kill_children=kill_children, killed=killed
+            )
+            return killed
+        return None
 
     # endregion
